mcp/requirements.txt
# MCP Gateway dependencies

# Web framework
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
starlette>=0.35.0,<1.0.0

# MCP SDK
mcp>=1.0.0,<2.0.0

# HTTP client
httpx>=0.27.0,<1.0.0

# Data validation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# JSON Schema validation
jsonschema>=4.20.0,<5.0.0

# YAML parsing
pyyaml>=6.0.0,<7.0.0

# Redis for caching
redis>=5.0.0,<6.0.0

# Logging
structlog>=24.1.0,<25.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0
respx>=0.20.0,<1.0.0
